# backend/ella/docs/hermes-cloud-runtime-targets.openapi.yaml
openapi: 3.1.0
info:
  title: Ella Hermes Cloud Runtime Targets
  version: 0.2.0
  description: >
    Backend-owned contract for account/profile Cloud runtime target resolution
    and AI processor consent. This fragment documents the Ella-specific surfaces;
    the generated public OMI OpenAPI file is not used for these internal routes.
servers:
  - url: https://api.ella-ai-care.com
security:
  - bearerAuth: []
paths:
...
  /v1/ella/onboarding/ensure:
    post:
      tags: [Runtime Provisioning]
      summary: Idempotently ensure the authenticated account/profile runtime.
      x-ella-runtime-target:
        mode: hermes-cloud-chat
        fallback: none
        sends-user-data: false
      responses:
        "200":
          description: Existing active exact target.
        "202":
          description: Provisioning or rollback remains in progress.
        "409":
          description: Current authority or immutable claim is invalid.
        "503":
          description: Schema, capacity, or exact target is unavailable.
  /v1/ella/onboarding/status:
    get:
      tags: [Runtime Provisioning]
      summary: Return the authenticated account/profile provisioning receipt.
      x-ella-runtime-target:
        mode: hermes-cloud-chat
        fallback: none
        sends-user-data: false
      responses:
        "200":
          description: Current redacted provisioning state.
        "404":
          description: Setup has not started.
        "503":
          description: Provisioning authority is unavailable.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    observerToken:
      type: apiKey
      in: header
      name: X-Ella-Observer-Token
...